pyproject.toml
[project]
name = "agent-swarm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.79",
    "python-dotenv>=1.2.2",
]
